implemented support for retesting existing submission(s)
This is helpful if the original test cases were incorrect.
Also, in an earlier bug, test results were not associated
with their corresponding submission receipt, so retesting
can create new test results for these submissions.

+++ CloudCoder/src/org/cloudcoder/app/shared/model/SubmissionReceipt.java
@@ -104,7 +104,7 @@ public class SubmissionReceipt implements Serializable, IContainsEvent {
 	/**
 	 * @return the lastEditEventId
 	 */
-	public long getLastEditEventId() {
+	public int getLastEditEventId() {
 		return lastEditEventId;
 	}
